reporting/streamlit/requirements.txt
# streamlit 
# matplotlib
# networkx
# google-cloud-secret-manager
# numpy
# pandas
# duckdb
streamlit
duckdb
google-cloud-secret-manager
google-cloud-aiplatform
vertexai
networkx
matplotlib
plotly
streamlit-lottie